+++ core/src/main/java/org/vertexium/util/StreamUtils.java
@@ -9,12 +9,27 @@ public class StreamUtils {
     private static final int DEFAULT_BUFFER_SIZE = 1024 * 4;
 
     public static long copy(InputStream input, OutputStream output) throws IOException {
+        return copy(input, output, 0, Long.MAX_VALUE);
+    }
+
+    public static long copy(InputStream input, OutputStream output, long offset, long limit) throws IOException {
         byte[] buffer = new byte[DEFAULT_BUFFER_SIZE];
         long count = 0;
         int n;
-        while (-1 != (n = input.read(buffer))) {
+        if (offset > 0) {
+            long skipResult = input.skip(offset);
+            if (skipResult < offset) {
+                return 0;
+            }
+        }
+        long len = limit;
+        while (-1 != (n = input.read(buffer, 0, (int) Math.min(buffer.length, len)))) {
             output.write(buffer, 0, n);
             count += n;
+            len -= n;
+            if (len == 0) {
+                break;
+            }
         }
         return count;
     }
@@ -25,7 +40,17 @@ public class StreamUtils {
         return out.toByteArray();
     }
 
+    public static byte[] toBytes(InputStream in, long offset, long limit) throws IOException {
+        ByteArrayOutputStream out = new ByteArrayOutputStream();
+        copy(in, out, offset, limit);
+        return out.toByteArray();
+    }
+
     public static String toString(InputStream in) throws IOException {
         return new String(toBytes(in));
+    }
+
+    public static String toString(InputStream in, long offset, long limit) throws IOException {
+        return new String(toBytes(in, offset, limit));
     }
 }